fix: Fix bug an

+++ src/main/java/org/ngphthinh/dto/response/statistics/RevenueDataResponse.java
@@ -1,9 +1,11 @@
 package org.ngphthinh.dto.response.statistics;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,5 +15,5 @@ import java.time.LocalDate;
 public class RevenueDataResponse {
 
     private LocalDate date;
-    private Long revenue;
+    private BigDecimal revenue;
 }
